requirements.txt
aiogram>=3.0.0
aiogram-dialog>=2.0.0
PyYAML==6.0
asyncpg==0.27.0
loguru==0.6.0
aerich==0.7.1
psycopg2-binary==2.9.5
SQLAlchemy==1.4.44
pytz~=2022.4
APScheduler==3.10.1
tortoise-orm==0.19.3
smsactivate~=1.5
requests~=2.31.0
aiohttp~=3.8.6
qrcode~=7.4.2
aioredis
uvicorn~=0.23.2
fastapi~=0.104.1
starlette~=0.27.0